[ws-daemon] Only handle k8s operations in controller
